+++ src/Events/RequestWasHandledEvent.php
@@ -19,21 +19,15 @@ use Fortuneglobe\IceHawk\Interfaces\ServesRequestInfo;
  */
 final class RequestWasHandledEvent implements ServesEventData
 {
-
-	/** @var ServesRequestInfo */
-	private $requestInfo;
-
 	/** @var ServesPostRequestData|ServesGetRequestData */
 	private $request;
 
 	/**
-	 * @param ServesRequestInfo $requestInfo
 	 * @param ServesRequestData $request
 	 */
-	public function __construct( ServesRequestInfo $requestInfo, ServesRequestData $request )
+	public function __construct( ServesRequestData $request )
 	{
-		$this->requestInfo = $requestInfo;
-		$this->request     = $request;
+		$this->request = $request;
 	}
 
 	/**
@@ -41,7 +35,7 @@ final class RequestWasHandledEvent implements ServesEventData
 	 */
 	public function getRequestInfo()
 	{
-		return $this->requestInfo;
+		return $this->request->getRequestInfo();
 	}
 
 	/**