+++ tests/Unit/DomainRequestHandlers/GetRequestHandlerTest.php
@@ -6,6 +6,7 @@
 
 namespace Fortuneglobe\IceHawk\Tests\Unit\DomainRequestHandlers;
 
+use Fortuneglobe\IceHawk\RequestInfo;
 use Fortuneglobe\IceHawk\Requests\GetRequest;
 use Fortuneglobe\IceHawk\Requests\PostRequest;
 use Fortuneglobe\IceHawk\Tests\Unit\Fixtures\TestGetRequestHandler;
@@ -17,14 +18,14 @@ class GetRequestHandlerTest extends \PHPUnit_Framework_TestCase
 	 */
 	public function testInjectingAPostRequestThrowsException()
 	{
-		$postRequest = new PostRequest( [ ], [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], [ ] );
 
 		new TestGetRequestHandler( $postRequest );
 	}
 
 	public function testCanHandleValidRequest()
 	{
-		$getRequest = new GetRequest( [ ] );
+		$getRequest = new GetRequest( RequestInfo::fromEnv(), [ ] );
 
 		$handler = new TestGetRequestHandler( $getRequest );
 		$handler->handleRequest();