+++ tests/Unit/Requests/GetRequestTest.php
@@ -5,16 +5,21 @@
 
 namespace Fortuneglobe\IceHawk\Tests\Unit\Requests;
 
+use Fortuneglobe\IceHawk\RequestInfo;
 use Fortuneglobe\IceHawk\Requests\GetRequest;
 
 class GetRequestTest extends \PHPUnit_Framework_TestCase
 {
 	/**
+	 * @param array  $getData
+	 * @param string $key
+	 * @param string $expectedValue
+	 *
 	 * @dataProvider requestDataProvider
 	 */
 	public function testCanGetRequestValueByKey( array $getData, $key, $expectedValue )
 	{
-		$getRequest = new GetRequest( $getData );
+		$getRequest = new GetRequest( RequestInfo::fromEnv(), $getData );
 
 		$this->assertEquals( $expectedValue, $getRequest->get( $key ) );
 	}
@@ -41,11 +46,14 @@ class GetRequestTest extends \PHPUnit_Framework_TestCase
 	}
 
 	/**
+	 * @param array  $getData
+	 * @param string $key
+	 *
 	 * @dataProvider nullKeyDataProvider
 	 */
 	public function testGetterReturnsNullIfKeyIsNotSet( array $getData, $key )
 	{
-		$getRequest = new GetRequest( $getData );
+		$getRequest = new GetRequest( RequestInfo::fromEnv(), $getData );
 
 		$this->assertNull( $getRequest->get( $key ) );
 	}