+++ tests/Unit/Requests/PostRequestTest.php
@@ -5,6 +5,7 @@
 
 namespace Fortuneglobe\IceHawk\Tests\Unit\Requests;
 
+use Fortuneglobe\IceHawk\RequestInfo;
 use Fortuneglobe\IceHawk\Requests\PostRequest;
 use Fortuneglobe\IceHawk\Requests\UploadedFile;
 use Fortuneglobe\IceHawk\Tests\Unit\Mocks\PhpStreamMock;
@@ -16,7 +17,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 	 */
 	public function testCanGetRequestValueByKey( array $postData, $key, $expectedValue )
 	{
-		$postRequest = new PostRequest( $postData, [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), $postData, [ ] );
 
 		$this->assertEquals( $expectedValue, $postRequest->get( $key ) );
 	}
@@ -47,7 +48,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 	 */
 	public function testGetterReturnsNullIfKeyIsNotSet( array $postData, $key )
 	{
-		$postRequest = new PostRequest( $postData, [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), $postData, [ ] );
 
 		$this->assertNull( $postRequest->get( $key ) );
 	}
@@ -79,7 +80,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 		$expectedError, $expectedErrorMessage, $expectedSuccess
 	)
 	{
-		$postRequest = new PostRequest( [ ], $uploadedFiles );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], $uploadedFiles );
 		$oneFile     = $postRequest->getOneFile( $fieldKey, $fileIndex );
 
 		$this->assertInstanceOf( UploadedFile::class, $oneFile );
@@ -163,7 +164,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 
 	public function testGetOneFileReturnNullIfKeyIsNotSet()
 	{
-		$postRequest = new PostRequest( [ ], [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], [ ] );
 		$oneFile     = $postRequest->getOneFile( 'test' );
 
 		$this->assertNull( $oneFile );
@@ -181,7 +182,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 			]
 		];
 
-		$postRequest = new PostRequest( [ ], $uploadedFiles );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], $uploadedFiles );
 		$oneFile     = $postRequest->getOneFile( 'test_file', 1 );
 
 		$this->assertNull( $oneFile );
@@ -189,7 +190,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 
 	public function testGetFilesReturnsEmptyArrayIfFieldKeyIsNotSet()
 	{
-		$postRequest = new PostRequest( [ ], [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], [ ] );
 		$files       = $postRequest->getFiles( 'test' );
 
 		$this->assertInternalType( 'array', $files );
@@ -202,7 +203,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 		stream_wrapper_register( "php", PhpStreamMock::class );
 		file_put_contents( 'php://input', 'Unit-Test' );
 
-		$postRequest = new PostRequest( [ ], [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], [ ] );
 
 		$this->assertEquals( 'Unit-Test', $postRequest->getRawData() );
 
@@ -211,7 +212,7 @@ class PostRequestTest extends \PHPUnit_Framework_TestCase
 
 	public function testGetRawDataReturnNullIfEmpty()
 	{
-		$postRequest = new PostRequest( [ ], [ ] );
+		$postRequest = new PostRequest( RequestInfo::fromEnv(), [ ], [ ] );
 
 		$this->assertNull( $postRequest->getRawData() );
 	}